feat(cmdsys): Introduce command processing system
This module is the foundation to our central location for processing
commands and results. This commit moves the logic for handling command
results out of the Game module and in to its own module.

This paves the way for us to move the remaining command logic out of
the Game module and into the cnetral command processing system.