refactor(game)!: Change Game singleton to pointer
This refactor converts the `Game` singleton (instance) to
'pass-by-pointer`. `Renderer`, `World`, and `Command System` are now created
in `main.c` and are passed to `game_create` to return a `Game` pointer,
which is passed to `game_init`, `game_update`, and `game_free`.

This commit also includes adding a unique ID to `Actor`. This is in
preperation for using a hash map.

Bulk of changes on this commit are clean-ups and restyling.

Note: `game_shutdown` has been renamed to `game_free`.

BREAKING CHANGE: `Game` singleton/instance has been removed. A `Game*`
handle is now passed around the main modules.
All call sites and usage has been reworked for the new API.